Initial Commit
Contains a basic webapp that lets you upload images or give a link and will then classify the collection of images. Right now this just calls Magic(), which should be replaced with the finished classifier. The recommender also does not call the finished recommender yet, and instead recommends a restaurant with the fewest reviews as a placeholder